.unpublished/ollama-refactor-code/files/pyproject.toml
[project]
name = "ai-reviewer"
version = "0.1.0"
description = "Elite AI Code Reviewer via Ollama"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.31.0",
    "pyyaml>=6.0.1",
]

[dependency-groups]
dev = [
    "ruff",
    "mypy",
    "types-requests",
    "types-pyyaml",
]

# INDISPENSABLE pour que "uv pip install ." fonctionne
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "UP", "B", "ANN"]

[tool.mypy]
strict = true